delete comment

+++ frontend/src/components/Comment.tsx
@@ -1,18 +1,20 @@
 "use client";
-import { IPost } from "@/lib/types";
+import { IComment, IPost } from "@/lib/types";
 import React from "react";
 import DOMPurify from "dompurify";
 import { Avatar, AvatarImage } from "./ui/avatar";
 import { Heart, MessageCircle, Share, Trash2 } from "lucide-react";
-import { deletePost } from "@/lib/utils";
+import { deleteComment, deletePost } from "@/lib/utils";
 import toast from "react-hot-toast";
 
 const Comment = ({
+  post,
   comment,
   userId,
   onDelete,
 }: {
-  comment: IPost;
+  post: IPost;
+  comment: IComment;
   userId: string | undefined;
   onDelete: () => void;
 }) => {
@@ -21,12 +23,19 @@ const Comment = ({
       __html: DOMPurify.sanitize(html),
     };
   };
+  console.log(comment);
 
   const handleDelete = async () => {
-    await deletePost(comment._id, userId!).then((res) => {
-      onDelete();
-      toast.success("comment deleted successfully");
-    });
+    if (userId !== undefined) {
+      await deleteComment(post._id, comment.commentId, userId)
+        .then((data) => {
+          onDelete();
+          toast.success("Comment deleted successfully");
+        })
+        .catch((error) => {
+          toast.error(error.message);
+        });
+    }
   };
 
   return (
@@ -44,23 +53,15 @@ const Comment = ({
       <div className="flex items-center justify-between w-full">
         <div className="flex items-center gap-4 mt-3">
           <button className="flex items-center gap-2">
-            <Heart size={18} color="#000" />
-            <h1>{comment.stats?.likes.count}</h1>
-          </button>
-          <button className="flex items-center gap-2">
             <MessageCircle size={18} color="#000" />
-            <h1>{comment.comments?.length}</h1>
+            <h1>{post.comments?.length}</h1>
           </button>
-          <button className="flex items-center gap-2">
-            <Share size={18} color="#000" />
-            <h1>{comment.stats?.shares}</h1>
-          </button>
+          {(userId === comment.userId || post.userId === userId) && (
+            <button className="flex items-center gap-2" onClick={handleDelete}>
+              <Trash2 size={18} color="#000" />
+            </button>
+          )}
         </div>
-        {userId === comment.userId && (
-          <button className="flex items-center gap-2" onClick={handleDelete}>
-            <Trash2 size={18} color="#000" />
-          </button>
-        )}
       </div>
     </div>
   );
